Final Summary — Second Deep Review
  You were right to push back. The second round used 5 specialized
cross-cutting auditors (concurrency, edge cases,
  conversion semantics, Matrices algorithms, equals/hashCode/Stream
contracts) targeting categories that per-file
  sibling comparison is blind to — bugs that exist equally across all
siblings.

  One genuine bug found and fixed:

  BooleanMatrix.countTrue() — silent integer overflow (severity:
low–medium, latent).

  - Root cause: Method declared public int countTrue() with int count =
0 accumulator. But
  AbstractMatrix.elementCount is a long and matrices may legally contain
more than Integer.MAX_VALUE cells (e.g.
  ~46341×46341), so a fully-true matrix would wrap to a negative result
with no exception.
  - Why per-file agents missed it: All sibling-comparison and
oracle-based techniques look for deviations between
  classes. This bug is unique to BooleanMatrix (no sibling has
countTrue), and javadoc-driven review found the code
  "matches its documentation" — except the documentation itself was
wrong.
  - Fix (BooleanMatrix.java:2110): return type and accumulator widened
from int to long; javadoc and example
  updated. Source-compatible for all 4 existing call sites (literal-int
comparisons widen to the
  assertEquals(long,long) overload).
  - Regression tests (4 new tests in
BooleanMatrixTest.CountTrueOverflowFix): reflection-based return-type
check
  that fails on the original int and passes on long, plus compile-time
and behavioral checks that the small-matrix
  and empty-matrix paths remain correct.

  Other auditors' results (all NO bugs found with concrete evidence):

  - Concurrency/parallel — traced every parallelization branch in
Matrices.forEachCartesianIndices and
  forEachIndices; the engine always parallelizes the disjoint axis (i or
j), never the accumulation axis k. matmul
  result[i][j] += ... is race-free in all 4 case branches across all 7
primitive matrix types.
  shareIntermediateArray is force-disabled when parallel in all 9 zip
sites. saturatedMultiply verified correct on
  every Long.MIN_VALUE/-1 edge case.
  - Edge cases/overflow — 26 adversarial scenarios mentally traced (0×0,
0×N, N×0, 1×1, 1×N, N×1, padding=0, equal
  copy bounds, null elements, –0.0 fill, MAX-sized SIZE products). All
correctly handled; the
  BooleanMatrix.countTrue overflow was the lone real finding.
  - Conversion semantics — 23 conversion edge cases traced through
actual JLS cast semantics (sign extension for
  byte/short→int/long, zero-extension for char→int, NaN/Infinity
saturation in (int)double and (long)double,
  precision loss past 2²⁴/2⁵³, round-trip soundness, overload binding).
All paths conform to spec and documentation.
  - Matrices algorithms — resolveCommonAssignableType BFS + 4-tier
comparator verified on 11 type pairs
  (Integer+Long → Number; ArrayList+LinkedList → List; int[]+Integer[] →
Cloneable; etc.); runWithParallelMode
  exception-safe; newMatrixArray correctly wraps primitives via
ClassUtil.wrap.
  - equals/hashCode/Stream — 28 contract-verification scenarios
concrete-traced (mixed-type equals, NaN/-0.0 bit
  semantics via N.deepEquals, NoSuchElementException on exhausted
iterators, count()/advance() using long
  arithmetic, Objectory recycled in finally).

  Final test run: 6477 tests, 0 failures, 0 errors. One real bug found,
fixed, and locked in by tests; thousands of
  lines of evidence that the rest of the codebase is correct.

+++ src/main/java/com/landawn/abacus/matrix/BooleanMatrix.java
@@ -1058,7 +1058,7 @@ public final class BooleanMatrix extends AbstractMatrix<boolean[], BooleanList, 
      * @param mapper the function to convert boolean values to type R
      * @param targetElementType the Class object for type R
      * @return a new Matrix containing the converted values
-     * @throws IllegalArgumentException if {@code mapper} is {@code null}
+     * @throws IllegalArgumentException if {@code mapper} or {@code targetElementType} is {@code null}
      * @throws E if the function throws an exception
      */
     public <R, E extends Exception> Matrix<R> mapToObj(final Throwables.BooleanFunction<? extends R, E> mapper, final Class<R> targetElementType) throws E {
@@ -1305,7 +1305,8 @@ public final class BooleanMatrix extends AbstractMatrix<boolean[], BooleanList, 
      * @param newRowCount the row count of the returned matrix; must be {@code >= 0}
      * @param newColumnCount the column count of the returned matrix; must be {@code >= 0}
      * @return a new BooleanMatrix with the specified dimensions
-     * @throws IllegalArgumentException if {@code newRowCount} or {@code newColumnCount} is negative
+     * @throws IllegalArgumentException if {@code newRowCount} or {@code newColumnCount} is negative,
+     *         or if {@code (long) newRowCount * newColumnCount} overflows {@code Integer.MAX_VALUE}
      * @see #resize(int, int, boolean)
      * @see #extend(int, int, int, int)
      */
@@ -1431,7 +1432,8 @@ public final class BooleanMatrix extends AbstractMatrix<boolean[], BooleanList, 
      * @param padLeft number of padding columns to add to the left of the original matrix; must be {@code >= 0}
      * @param padRight number of padding columns to add to the right of the original matrix; must be {@code >= 0}
      * @return a new BooleanMatrix with dimensions {@code (padTop + rowCount + padBottom) × (padLeft + columnCount + padRight)}
-     * @throws IllegalArgumentException if any parameter is negative
+     * @throws IllegalArgumentException if any parameter is negative,
+     *         or if the resulting dimensions would overflow {@code Integer.MAX_VALUE}
      * @see #extend(int, int, int, int, boolean)
      * @see #resize(int, int)
      */
@@ -2097,16 +2099,21 @@ public final class BooleanMatrix extends AbstractMatrix<boolean[], BooleanList, 
     /**
      * Counts the number of {@code true} elements in this matrix.
      *
+     * <p>The result is a {@code long} because a matrix can contain up to
+     * {@code rowCount * columnCount} elements, a product that may exceed
+     * {@link Integer#MAX_VALUE}. An {@code int} accumulator would silently
+     * overflow for sufficiently large matrices.</p>
+     *
      * <p><b>Usage Examples:</b></p>
      * <pre>{@code
      * BooleanMatrix matrix = BooleanMatrix.of(new boolean[][] {{true, false, true}, {false, true, false}});
-     * int count = matrix.countTrue();   // Returns 3
+     * long count = matrix.countTrue();   // Returns 3
      * }</pre>
      *
-     * @return the number of {@code true} elements in this matrix
+     * @return the number of {@code true} elements in this matrix, as a non-negative {@code long}
      */
-    public int countTrue() {
-        int count = 0;
+    public long countTrue() {
+        long count = 0;
 
         for (int i = 0; i < rowCount; i++) {
             final boolean[] row = a[i];
